+++ tests/test_table.py
@@ -8,12 +8,18 @@ def test_table(document):
     elem.set_headers(['one', 'two', 'three'])
     elem.add_row(['Hello', 'World', '!'])
     elem.add_row(['foo', 'bar', 64])
+    elem.add_rows([
+        ['multiple', 'rows', 'test 1'],
+        ['multiple', 'rows', 'test 2'],
+    ])
     document.add(elem)
 
-    expected = '| one   | two   | three |\n' \
-               '| ----- | ----- | ----- |\n' \
-               '| Hello | World | !     |\n' \
-               '| foo   | bar   | 64    |\n\n'
+    expected = '| one      | two   | three  |\n' \
+               '| -------- | ----- | ------ |\n' \
+               '| Hello    | World | !      |\n' \
+               '| foo      | bar   | 64     |\n' \
+               '| multiple | rows  | test 1 |\n' \
+               '| multiple | rows  | test 2 |\n\n'
     assert expected == document.output()