+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.3 (--)
 
 * Blockquote block can handle universal newlines.
+* Multiple rows at a time can be added in table.
 
 ## 0.0.2 (2020-07-02)